do not use null conditional invoke, we don't need it
references are being kept alive by the event bus itself. this means the methods do not get garbage collected even if the class is gone (including Destroy of MonoBehaviour). people should call RemoveListener anyway, so this is not a big issue. still frustrating how c# doesn't have a WeakDelegate though.