optimization: stackalloc
This one did not improve performance much, but it reduced GC allocs drastically because we are not allocating anything on the heap when passing in channel arrays for multi-channel emits. 

It is kind of cumbersome and we had to make the EventMetadata a ref struct (meaning it can only exist on the stack), but I think it is worth it overall.